Set ls shortcuts in pwsh same as in bash

+++ powershell/Microsoft.PowerShell_profile.ps1
@@ -16,13 +16,8 @@ function which {
     Get-Command $command | Select-Object -ExpandProperty Definition
 }
 
-function l { 
-    & ls.exe -hAlF 
-}
-
-Set-Alias ll l
-
-if (Test-Path Alias:ls) {
-    Remove-Item Alias:ls
-}
+# Set ls shortcuts the same as in bash/zsh:
+function l { & ls.exe -XF }
+function ll { & ls.exe -hAlF }
+if (Test-Path Alias:ls) { Remove-Item Alias:ls }
